修改第一题

+++ 1.c
@@ -1,25 +1,28 @@
 #include<stdio.h>
 int main(){
 
-	int a, b = 0;
+	long long a, b = 0;
 	int arr[10];
 	int flag = 1;
-	scanf("%d", &a);
+	scanf("%lld", &a);
 	flag = a > 0 ? 1 : -1;
 	for(int i=1; i <=9; i++){
 		scanf("%d", &arr[i]);
 	}
-	int t = a * flag;
-	int n = 1; 
+	//printf("a = %lld\n", a);
+	long long t = a * flag;
+	long long n = 1; 
 	while(t > n){
 		n = n * 10;
 	}
+	//printf("n = %d\n", n);
 	n = n / 10; 
 
 	while(n > 0){
+		//printf("b = %lld, t = %lld, n = %d\n", b , t, n);
 		b = b + arr[t/n] * n;
 		t = t - t/n * n;
 		n = n / 10; 
 	}
-	printf("%d\n", b * flag);
+	printf("%lld\n", b * flag);
 }
